Extract run blob I/O into BlobService; split audit handler
Refactor the oversized pipeline_runs handler and its storage orchestration:

- Add BlobService (service/blob.rs), owning the NatsClient + CryptoService,
  with the run-blob helpers (build_document, store_redacted_file,
  store_analyzed_document, load_analyzed_document). Wired via ServiceState DI.
  The handlers now call blob.* and drop the nats/crypto State they no longer
  use directly.
- Move the engine-error mapping to a From<nvisy_engine::Error> impl in
  handler/error/engine_error.rs, matching the pg/crypto/nats/object pattern;
  analyze/anonymize sites become plain `?` / `.into()`.
- Split the run's read/export endpoints (/detections/, /audit/json,
  /audit/csv) into a sibling pipeline_audits handler, sharing find_pipeline_run
  (now pub(super)). Collapse the CSV-zip error closures into one local
  archive_error helper.
- Flatten the blob/ and object/ service module directories to single files.

pipeline_runs.rs drops from ~1100 to ~640 lines. No behavior change.
