+++ crates/nvisy-server/src/handler/error/mod.rs
@@ -1,6 +1,7 @@
 //! [`Error`], [`ErrorKind`] and [`Result`].
 
 mod crypto_error;
+mod engine_error;
 mod http_error;
 mod nats_error;
 mod object_error;
