+++ crates/nvisy-server/src/handler/pipeline_runs.rs
@@ -4,37 +4,22 @@
 //! and stores the findings; the run then awaits reviewer verification before
 //! redact consumes the verified findings and produces a redacted file.
 
-use std::io::{Cursor, Write as _};
-use std::str::FromStr;
-
 use aide::axum::ApiRouter;
 use aide::transform::TransformOperation;
-use axum::body::Body;
 use axum::extract::State;
-use axum::http::header::{CONTENT_DISPOSITION, CONTENT_LENGTH, CONTENT_TYPE};
-use axum::http::{HeaderMap, HeaderValue, StatusCode};
-use bytes::Bytes;
+use axum::http::StatusCode;
+use nvisy_engine::OcrMode;
 use nvisy_engine::policy::PolicyDefinition;
-use nvisy_engine::{Audit, Document, OcrMode};
-use nvisy_nats::NatsClient;
-use nvisy_nats::object::{AuditBucket, AuditKey, FileKey, FilesBucket};
 use nvisy_postgres::model::{
-    NewWorkspaceFile, NewWorkspacePipelineRun, UpdateWorkspacePipelineRun, WorkspaceFile,
-    WorkspacePipeline, WorkspacePipelineRun,
+    NewWorkspacePipelineRun, UpdateWorkspacePipelineRun, WorkspacePipeline, WorkspacePipelineRun,
 };
 use nvisy_postgres::query::{
     PipelineReferenceRepository, WorkspaceFileRepository, WorkspacePipelineRepository,
     WorkspacePipelineRunRepository, WorkspacePolicyRepository,
 };
-use nvisy_postgres::types::{
-    FileKind, OcrPolicy, PipelineRunStatus, RetentionOverride, RetentionScope, RetentionSettings,
-    WorkspaceSettings,
-};
+use nvisy_postgres::types::{OcrPolicy, PipelineRunStatus, WorkspaceSettings};
 use nvisy_postgres::{PgClient, PgConn};
-use sha2::{Digest, Sha256};
-use tokio::io::AsyncReadExt;
 use uuid::Uuid;
-use zip::write::SimpleFileOptions;
 
 use crate::extract::{
     AuthProvider, AuthState, IdempotencyKey, Json, Path, Permission, Query, ValidateJson,
@@ -47,7 +32,7 @@ use crate::handler::request::{
 use crate::handler::response::{ErrorResponse, PipelineRun, PipelineRunsPage};
 use crate::handler::utility::resolve_account_ref;
 use crate::handler::{Error, ErrorKind, Result};
-use crate::service::{CryptoService, EngineService, ServiceState, WebhookEmitter};
+use crate::service::{BlobService, CryptoService, EngineService, ServiceState, WebhookEmitter};
 
 /// Tracing target for pipeline run operations.
 const TRACING_TARGET: &str = "nvisy_server::handler::runs";
@@ -67,7 +52,7 @@ const TRACING_TARGET: &str = "nvisy_server::handler::runs";
 )]
 async fn create_pipeline_run(
     State(pg_client): State<PgClient>,
-    State(nats): State<NatsClient>,
+    State(blob): State<BlobService>,
     State(crypto): State<CryptoService>,
     State(engine): State<EngineService>,
     State(webhook_emitter): State<WebhookEmitter>,
@@ -155,7 +140,7 @@ async fn create_pipeline_run(
     // recognizer lineups and enrichers are engine-owned.
     let params = engine.analyzer_params(&definition, request.scope, ocr_mode);
 
-    let document = build_document(&nats, &crypto, &file, run.id).await?;
+    let document = blob.build_document(&file, run.id).await?;
 
     // The pipeline's policies own the label vocabulary: detection derives its
     // catalog from them, so the analysis emits exactly what the policies can act
@@ -186,23 +171,22 @@ async fn create_pipeline_run(
                 auth_state.account_id,
             )
             .await;
-            return Err(analysis_error(err));
+            return Err(err.into());
         }
     };
 
     // The analysis is a map of detected PII; encrypt it and record it as an
     // audit-kind file, keeping only its id on the run.
     let workspace_settings = WorkspaceSettings::from_value(&workspace.settings).retention;
-    let audit_file_id = store_analyzed_document(
-        &mut conn,
-        &nats,
-        &crypto,
-        &pipeline,
-        &workspace_settings,
-        auth_state.account_id,
-        &analyzed,
-    )
-    .await?;
+    let audit_file_id = blob
+        .store_analyzed_document(
+            &mut conn,
+            &pipeline,
+            &workspace_settings,
+            auth_state.account_id,
+            &analyzed,
+        )
+        .await?;
     let run = conn
         .update_workspace_pipeline_run(
             run.id,
@@ -413,224 +397,6 @@ fn get_pipeline_run_docs(op: TransformOperation) -> TransformOperation {
         .response::<404, Json<ErrorResponse>>()
 }
 
-/// Returns the run's analysis (the detected findings) for review.
-///
-/// Fetches and decrypts the engine's `Audit` from the intermediates bucket.
-/// Available once the run is analyzed. Requires `ViewPipelines`.
-#[tracing::instrument(
-    skip_all,
-    fields(
-        account_id = %auth_state.account_id,
-        workspace_id = %workspace.id,
-        run_id = %path_params.run_id,
-    )
-)]
-async fn get_pipeline_run_analysis(
-    State(pg_client): State<PgClient>,
-    State(nats): State<NatsClient>,
-    State(crypto): State<CryptoService>,
-    AuthState(auth_state): AuthState,
-    WorkspaceContext(workspace): WorkspaceContext,
-    Path(path_params): Path<PipelineRunPathParams>,
-) -> Result<(StatusCode, Json<Audit>)> {
-    tracing::debug!(target: TRACING_TARGET, "Getting pipeline run analysis");
-
-    let mut conn = pg_client.get_connection().await?;
-
-    auth_state
-        .authorize_workspace(&mut conn, workspace.id, Permission::ViewPipelines)
-        .await?;
-
-    let (run, _pipeline) =
-        find_pipeline_run(&mut conn, workspace.id, path_params.run_id.as_uuid()).await?;
-
-    let analyzed = load_analyzed_document(&mut conn, &nats, &crypto, workspace.id, &run).await?;
-
-    tracing::debug!(target: TRACING_TARGET, "Pipeline run analysis retrieved");
-
-    Ok((StatusCode::OK, Json(analyzed)))
-}
-
-fn get_pipeline_run_analysis_docs(op: TransformOperation) -> TransformOperation {
-    op.summary("Get run detections")
-        .description("Returns the run's detected findings (the analyzed document) for review.")
-        .response::<200, Json<Audit>>()
-        .response::<401, Json<ErrorResponse>>()
-        .response::<403, Json<ErrorResponse>>()
-        .response::<404, Json<ErrorResponse>>()
-        .response::<409, Json<ErrorResponse>>()
-}
-
-/// Downloads a run's audit as a pretty-printed JSON file.
-///
-/// The full structure — body, parts, context, and each entity's provenance
-/// chain. Requires `ViewPipelines` permission.
-#[tracing::instrument(
-    skip_all,
-    fields(
-        account_id = %auth_state.account_id,
-        workspace_id = %workspace.id,
-        run_id = %path_params.run_id,
-    )
-)]
-async fn download_pipeline_run_audit_json(
-    State(pg_client): State<PgClient>,
-    State(nats): State<NatsClient>,
-    State(crypto): State<CryptoService>,
-    AuthState(auth_state): AuthState,
-    WorkspaceContext(workspace): WorkspaceContext,
-    Path(path_params): Path<PipelineRunPathParams>,
-) -> Result<(StatusCode, HeaderMap, Body)> {
-    tracing::debug!(target: TRACING_TARGET, "Downloading pipeline run audit (JSON)");
-
-    let mut conn = pg_client.get_connection().await?;
-    auth_state
-        .authorize_workspace(&mut conn, workspace.id, Permission::ViewPipelines)
-        .await?;
-
-    let (run, _pipeline) =
-        find_pipeline_run(&mut conn, workspace.id, path_params.run_id.as_uuid()).await?;
-    let audit = load_analyzed_document(&mut conn, &nats, &crypto, workspace.id, &run).await?;
-
-    let mut buffer = Vec::new();
-    audit.write_json(&mut buffer).map_err(|err| {
-        ErrorKind::InternalServerError
-            .with_message("Failed to export audit as JSON")
-            .with_context(err.to_string())
-    })?;
-
-    let filename = format!("audit-{}.json", run.id);
-    let headers = attachment_headers(
-        &filename,
-        HeaderValue::from_static("application/json"),
-        buffer.len(),
-    );
-
-    tracing::debug!(target: TRACING_TARGET, "Pipeline run audit exported (JSON)");
-    Ok((StatusCode::OK, headers, Body::from(buffer)))
-}
-
-fn download_pipeline_run_audit_json_docs(op: TransformOperation) -> TransformOperation {
-    op.summary("Download run audit (JSON)")
-        .description("Downloads the run's audit as a pretty-printed JSON file.")
-        .response::<200, ()>()
-        .response::<401, Json<ErrorResponse>>()
-        .response::<403, Json<ErrorResponse>>()
-        .response::<404, Json<ErrorResponse>>()
-        .response::<409, Json<ErrorResponse>>()
-}
-
-/// Downloads a run's audit as a zip of CSV tables.
-///
-/// The archive holds `entities.csv`, `provenance.csv`, and `reviews.csv`, which
-/// join on `entity_id`. Requires `ViewPipelines` permission.
-#[tracing::instrument(
-    skip_all,
-    fields(
-        account_id = %auth_state.account_id,
-        workspace_id = %workspace.id,
-        run_id = %path_params.run_id,
-    )
-)]
-async fn download_pipeline_run_audit_csv(
-    State(pg_client): State<PgClient>,
-    State(nats): State<NatsClient>,
-    State(crypto): State<CryptoService>,
-    AuthState(auth_state): AuthState,
-    WorkspaceContext(workspace): WorkspaceContext,
-    Path(path_params): Path<PipelineRunPathParams>,
-) -> Result<(StatusCode, HeaderMap, Body)> {
-    tracing::debug!(target: TRACING_TARGET, "Downloading pipeline run audit (CSV)");
-
-    let mut conn = pg_client.get_connection().await?;
-    auth_state
-        .authorize_workspace(&mut conn, workspace.id, Permission::ViewPipelines)
-        .await?;
-
-    let (run, _pipeline) =
-        find_pipeline_run(&mut conn, workspace.id, path_params.run_id.as_uuid()).await?;
-    let audit = load_analyzed_document(&mut conn, &nats, &crypto, workspace.id, &run).await?;
-
-    let archive = build_audit_csv_zip(&audit)?;
-
-    let filename = format!("audit-{}.csv.zip", run.id);
-    let headers = attachment_headers(
-        &filename,
-        HeaderValue::from_static("application/zip"),
-        archive.len(),
-    );
-
-    tracing::debug!(target: TRACING_TARGET, "Pipeline run audit exported (CSV)");
-    Ok((StatusCode::OK, headers, Body::from(archive)))
-}
-
-fn download_pipeline_run_audit_csv_docs(op: TransformOperation) -> TransformOperation {
-    op.summary("Download run audit (CSV)")
-        .description(
-            "Downloads the run's audit as a zip of entities.csv, provenance.csv, and reviews.csv.",
-        )
-        .response::<200, ()>()
-        .response::<401, Json<ErrorResponse>>()
-        .response::<403, Json<ErrorResponse>>()
-        .response::<404, Json<ErrorResponse>>()
-        .response::<409, Json<ErrorResponse>>()
-}
-
-/// Bundles the audit's three CSV tables into an in-memory zip archive.
-fn build_audit_csv_zip(audit: &Audit) -> Result<Vec<u8>> {
-    let map_csv = |err: nvisy_engine::Error| {
-        ErrorKind::InternalServerError
-            .with_message("Failed to export audit as CSV")
-            .with_context(err.to_string())
-    };
-    let archive_error = |err: &dyn std::fmt::Display| {
-        ErrorKind::InternalServerError
-            .with_message("Failed to build audit archive")
-            .with_context(err.to_string())
-    };
-    let map_zip = |err: zip::result::ZipError| archive_error(&err);
-    let map_io = |err: std::io::Error| archive_error(&err);
-
-    let mut writer = zip::ZipWriter::new(std::io::Cursor::new(Vec::new()));
-    let options = SimpleFileOptions::default();
-
-    let mut table = Vec::new();
-    audit.write_entities_csv(&mut table).map_err(map_csv)?;
-    writer
-        .start_file("entities.csv", options)
-        .map_err(map_zip)?;
-    writer.write_all(&table).map_err(map_io)?;
-
-    table.clear();
-    audit.write_provenance_csv(&mut table).map_err(map_csv)?;
-    writer
-        .start_file("provenance.csv", options)
-        .map_err(map_zip)?;
-    writer.write_all(&table).map_err(map_io)?;
-
-    table.clear();
-    audit.write_reviews_csv(&mut table).map_err(map_csv)?;
-    writer.start_file("reviews.csv", options).map_err(map_zip)?;
-    writer.write_all(&table).map_err(map_io)?;
-
-    Ok(writer.finish().map_err(map_zip)?.into_inner())
-}
-
-/// Builds the download response headers for an attachment.
-///
-/// `filename` is server-generated (a run UUID plus a fixed extension), so it
-/// needs no sanitization for the quoted header value.
-fn attachment_headers(filename: &str, content_type: HeaderValue, length: usize) -> HeaderMap {
-    let mut headers = HeaderMap::new();
-    let disposition = format!("attachment; filename=\"{filename}\"")
-        .parse()
-        .unwrap_or_else(|_| HeaderValue::from_static("attachment"));
-    headers.insert(CONTENT_DISPOSITION, disposition);
-    headers.insert(CONTENT_TYPE, content_type);
-    headers.insert(CONTENT_LENGTH, HeaderValue::from(length as u64));
-    headers
-}
-
 /// Redacts a run using the reviewer-verified findings, storing the result.
 ///
 /// Consumes the analyzed run (which must be awaiting review), applies the
@@ -646,7 +412,7 @@ fn attachment_headers(filename: &str, content_type: HeaderValue, length: usize) 
 )]
 async fn redact_pipeline_run(
     State(pg_client): State<PgClient>,
-    State(nats): State<NatsClient>,
+    State(blob): State<BlobService>,
     State(crypto): State<CryptoService>,
     State(engine): State<EngineService>,
     State(webhook_emitter): State<WebhookEmitter>,
@@ -687,29 +453,26 @@ async fn redact_pipeline_run(
     // The stored analysis is the source of truth for what gets redacted. It
     // carries the scope (label catalog) analyze resolved, so redaction compiles
     // against the same vocabulary without re-deriving it.
-    let mut analyzed =
-        load_analyzed_document(&mut conn, &nats, &crypto, workspace.id, &run).await?;
+    let mut analyzed = blob
+        .load_analyzed_document(&mut conn, workspace.id, &run)
+        .await?;
     let policies = resolve_policies(&mut conn, &crypto, workspace.id, pipeline.id).await?;
-    let document = build_document(&nats, &crypto, &file, run.id).await?;
+    let document = blob.build_document(&file, run.id).await?;
 
-    let redacted = engine
-        .anonymize(document, &policies, &mut analyzed)
-        .await
-        .map_err(analysis_error)?;
+    let redacted = engine.anonymize(document, &policies, &mut analyzed).await?;
 
     // Store the redacted bytes as a new workspace file and link it to the run.
     let workspace_settings = WorkspaceSettings::from_value(&workspace.settings).retention;
-    let output_file = store_redacted_file(
-        &mut conn,
-        &nats,
-        &crypto,
-        &file,
-        &pipeline,
-        &workspace_settings,
-        auth_state.account_id,
-        redacted.bytes,
-    )
-    .await?;
+    let output_file = blob
+        .store_redacted_file(
+            &mut conn,
+            &file,
+            &pipeline,
+            &workspace_settings,
+            auth_state.account_id,
+            redacted.bytes,
+        )
+        .await?;
 
     let run = conn
         .update_workspace_pipeline_run(
@@ -811,20 +574,6 @@ fn serialize_error(error: serde_json::Error) -> Error<'static> {
         .with_context(error.to_string())
 }
 
-/// Maps an analyzed-document (de)serialization failure to an internal error.
-fn analysis_serde_error(error: serde_json::Error) -> Error<'static> {
-    ErrorKind::InternalServerError
-        .with_message("Failed to process analysis")
-        .with_context(error.to_string())
-}
-
-/// Maps an engine analyze/anonymize failure to an internal error.
-fn analysis_error(error: nvisy_engine::Error) -> Error<'static> {
-    ErrorKind::InternalServerError
-        .with_message("Redaction engine failed")
-        .with_context(error.to_string())
-}
-
 /// Finds a pipeline within a workspace by slug or returns NotFound.
 async fn find_pipeline(
     conn: &mut PgConn,
@@ -840,7 +589,7 @@ async fn find_pipeline(
 /// Resolves a run by its opaque id within a workspace, returning the run and its
 /// owning pipeline (for the response's pipeline slug). The lookup is
 /// workspace-scoped through the owning pipeline.
-async fn find_pipeline_run(
+pub(super) async fn find_pipeline_run(
     conn: &mut PgConn,
     workspace_id: Uuid,
     run_id: Uuid,
@@ -871,65 +620,10 @@ pub fn routes() -> ApiRouter<ServiceState> {
             get_with(get_pipeline_run, get_pipeline_run_docs),
         )
         .api_route(
-            "/workspaces/{workspaceSlug}/runs/{runId}/detections/",
-            get_with(get_pipeline_run_analysis, get_pipeline_run_analysis_docs),
-        )
-        .api_route(
-            "/workspaces/{workspaceSlug}/runs/{runId}/audit/json",
-            get_with(
-                download_pipeline_run_audit_json,
-                download_pipeline_run_audit_json_docs,
-            ),
-        )
-        .api_route(
-            "/workspaces/{workspaceSlug}/runs/{runId}/audit/csv",
-            get_with(
-                download_pipeline_run_audit_csv,
-                download_pipeline_run_audit_csv_docs,
-            ),
-        )
-        .api_route(
             "/workspaces/{workspaceSlug}/runs/{runId}/redactions/",
             post_with(redact_pipeline_run, redact_pipeline_run_docs),
         )
         .with_path_items(|item| item.tag("Pipeline Runs"))
-}
-
-/// Reads a workspace file's bytes from object storage and builds an engine
-/// [`Document`], stamping the run's id as the correlation id.
-async fn build_document(
-    nats: &NatsClient,
-    crypto: &CryptoService,
-    file: &WorkspaceFile,
-    correlation_id: Uuid,
-) -> Result<Document> {
-    let store = nats.object_store::<FilesBucket>().await?;
-    let key = FileKey::from_str(&file.storage_path).map_err(|err| {
-        ErrorKind::InternalServerError
-            .with_message("Invalid file storage path")
-            .with_context(err.to_string())
-    })?;
-
-    let data = store.get(&key).await?.ok_or_else(|| {
-        ErrorKind::InternalServerError.with_message("File content is missing from storage")
-    })?;
-    let mut reader = data.into_reader();
-    let mut ciphertext = Vec::new();
-    reader.read_to_end(&mut ciphertext).await.map_err(|err| {
-        ErrorKind::InternalServerError
-            .with_message("Failed to read file content")
-            .with_context(err.to_string())
-    })?;
-
-    let bytes = crypto
-        .decrypt(file.workspace_id, &ciphertext)
-        .map_err(|err| {
-            ErrorKind::InternalServerError
-                .with_message("Failed to decrypt file content")
-                .with_context(err.to_string())
-        })?;
-
-    Ok(Document::new(bytes, file.file_extension.clone()).with_correlation_id(correlation_id))
 }
 
 /// Resolves a pipeline's live policy references into decrypted engine policies.
@@ -948,162 +642,4 @@ async fn resolve_policies(
         }
     }
     Ok(policies)
-}
-
-/// Stores redacted bytes as a new workspace file (the run's output).
-///
-/// The redacted file is a first-class file — a sibling of the source — so it is
-/// downloadable through the normal file endpoints.
-#[allow(clippy::too_many_arguments)]
-async fn store_redacted_file(
-    conn: &mut PgConn,
-    nats: &NatsClient,
-    crypto: &CryptoService,
-    source: &WorkspaceFile,
-    pipeline: &WorkspacePipeline,
-    workspace_settings: &RetentionSettings,
-    account_id: Uuid,
-    bytes: Bytes,
-) -> Result<WorkspaceFile> {
-    // Record the plaintext size and hash before encrypting; storage holds only
-    // the ciphertext.
-    let plaintext_size = bytes.len() as i64;
-    let plaintext_hash = Sha256::digest(&bytes).to_vec();
-    let ciphertext = crypto.encrypt(source.workspace_id, &bytes).map_err(|err| {
-        ErrorKind::InternalServerError
-            .with_message("Failed to encrypt redacted file")
-            .with_context(err.to_string())
-    })?;
-
-    let store = nats.object_store::<FilesBucket>().await?;
-    let key = FileKey::generate(source.workspace_id);
-    store.put(&key, Cursor::new(ciphertext)).await?;
-
-    // Retention expiry for the redacted-documents scope (workspace baseline,
-    // pipeline override if set).
-    let over = RetentionOverride::from_pipeline_metadata(&pipeline.metadata);
-    let expires_at = workspace_settings
-        .resolve(RetentionScope::RedactedDocuments, over.as_ref())
-        .expires_at(jiff::Timestamp::now());
-
-    let redacted_name = format!("{}.redacted", source.display_name);
-    let new_file = NewWorkspaceFile {
-        workspace_id: source.workspace_id,
-        account_id,
-        parent_id: Some(source.id),
-        display_name: Some(redacted_name),
-        original_filename: Some(source.original_filename.clone()),
-        file_extension: Some(source.file_extension.clone()),
-        file_kind: Some(FileKind::Redacted),
-        file_size_bytes: plaintext_size,
-        file_hash_sha256: plaintext_hash,
-        storage_path: key.to_string(),
-        storage_bucket: store.bucket().to_owned(),
-        expires_at: expires_at.map(Into::into),
-        ..Default::default()
-    };
-
-    Ok(conn.create_workspace_file(new_file).await?)
-}
-
-/// Encrypts an [`Audit`], stores it in the audit bucket, and records it as an
-/// `audit`-kind [`WorkspaceFile`], returning that file's id.
-///
-/// The analysis is the map of detected PII, so it is encrypted with the
-/// workspace key before it leaves the process. Modeling it as a file lets
-/// data-retention expire it with the same `expires_at` sweep as documents; its
-/// bytes live in the audit bucket, not the files bucket.
-async fn store_analyzed_document(
-    conn: &mut PgConn,
-    nats: &NatsClient,
-    crypto: &CryptoService,
-    pipeline: &WorkspacePipeline,
-    workspace_settings: &RetentionSettings,
-    account_id: Uuid,
-    analyzed: &Audit,
-) -> Result<Uuid> {
-    let workspace_id = pipeline.workspace_id;
-    let plaintext = serde_json::to_vec(analyzed).map_err(analysis_serde_error)?;
-    let hash = Sha256::digest(&plaintext).to_vec();
-    let size = plaintext.len() as i64;
-    let ciphertext = crypto.encrypt(workspace_id, &plaintext).map_err(|err| {
-        ErrorKind::InternalServerError
-            .with_message("Failed to encrypt analysis")
-            .with_context(err.to_string())
-    })?;
-
-    let store = nats.object_store::<AuditBucket>().await?;
-    let key = AuditKey::generate(workspace_id);
-    store.put(&key, Cursor::new(ciphertext)).await?;
-
-    // Retention expiry for the audit scope (workspace baseline, pipeline
-    // override if set).
-    let over = RetentionOverride::from_pipeline_metadata(&pipeline.metadata);
-    let expires_at = workspace_settings
-        .resolve(RetentionScope::AuditLogs, over.as_ref())
-        .expires_at(jiff::Timestamp::now());
-
-    let new_file = NewWorkspaceFile {
-        workspace_id,
-        account_id,
-        display_name: Some("analysis.audit".to_owned()),
-        original_filename: Some("analysis.audit".to_owned()),
-        file_extension: Some("json".to_owned()),
-        file_kind: Some(FileKind::Audit),
-        file_size_bytes: size,
-        file_hash_sha256: hash,
-        storage_path: key.to_string(),
-        storage_bucket: store.bucket().to_owned(),
-        expires_at: expires_at.map(Into::into),
-        ..Default::default()
-    };
-    let file = conn.create_workspace_file(new_file).await?;
-    Ok(file.id)
-}
-
-/// Fetches and decrypts a run's stored [`Audit`].
-///
-/// Errors if the run has not been analyzed yet or the stored object is missing.
-async fn load_analyzed_document(
-    conn: &mut PgConn,
-    nats: &NatsClient,
-    crypto: &CryptoService,
-    workspace_id: Uuid,
-    run: &WorkspacePipelineRun,
-) -> Result<Audit> {
-    let audit_file_id = run.audit_file_id.ok_or_else(|| {
-        ErrorKind::Conflict
-            .with_message("Run has no analysis yet")
-            .with_resource("pipeline_run")
-    })?;
-    let audit_file = conn
-        .find_file_in_workspace(workspace_id, audit_file_id)
-        .await?
-        .ok_or_else(|| {
-            ErrorKind::InternalServerError.with_message("Analysis is missing from storage")
-        })?;
-    let key = AuditKey::from_str(&audit_file.storage_path).map_err(|err| {
-        ErrorKind::InternalServerError
-            .with_message("Invalid analysis storage key")
-            .with_context(err.to_string())
-    })?;
-
-    let store = nats.object_store::<AuditBucket>().await?;
-    let data = store.get(&key).await?.ok_or_else(|| {
-        ErrorKind::InternalServerError.with_message("Analysis is missing from storage")
-    })?;
-    let mut reader = data.into_reader();
-    let mut ciphertext = Vec::new();
-    reader.read_to_end(&mut ciphertext).await.map_err(|err| {
-        ErrorKind::InternalServerError
-            .with_message("Failed to read analysis")
-            .with_context(err.to_string())
-    })?;
-
-    let plaintext = crypto.decrypt(workspace_id, &ciphertext).map_err(|err| {
-        ErrorKind::InternalServerError
-            .with_message("Failed to decrypt analysis")
-            .with_context(err.to_string())
-    })?;
-    serde_json::from_slice(&plaintext).map_err(analysis_serde_error)
 }
