+++ crates/nvisy-server/src/service/mod.rs
@@ -1,6 +1,7 @@
 //! Application state and dependency injection.
 
 mod avatar;
+mod blob;
 mod connection;
 pub mod crypto;
 pub mod engine;
@@ -19,6 +20,7 @@ use nvisy_postgres::{PgClient, PgClientMigrationExt, PgConfig};
 use nvisy_webhook::WebhookService;
 
 pub use crate::service::avatar::{AVATAR_CONTENT_TYPE, AvatarService, MAX_AVATAR_UPLOAD_BYTES};
+pub use crate::service::blob::BlobService;
 pub use crate::service::connection::ConnectionConfig;
 pub use crate::service::crypto::{CryptoConfig, CryptoService};
 pub(crate) use crate::service::crypto::{HashingReader, Measurements};
@@ -57,6 +59,7 @@ pub struct ServiceState {
 
     // Internal services:
     pub avatar: AvatarService,
+    pub blob: BlobService,
     pub connection_sync: ConnectionSyncService,
     pub health_cache: HealthCache,
     pub object: ObjectService,
@@ -95,6 +98,7 @@ impl ServiceState {
         ];
 
         let object = ObjectService::new();
+        let blob = BlobService::new(nats_client.clone(), crypto.clone());
         let avatar = AvatarService::new(nats_client.clone(), postgres_client.clone());
         let connection_sync = ConnectionSyncService::new(
             postgres_client.clone(),
@@ -114,6 +118,7 @@ impl ServiceState {
             engine,
 
             avatar,
+            blob,
             connection_sync,
             health_cache: HealthCache::new(&health_config, health_checkers),
             object,
@@ -192,6 +197,7 @@ impl_di!(
 // Internal services:
 impl_di!(
     avatar: AvatarService,
+    blob: BlobService,
     connection_sync: ConnectionSyncService,
     crypto: CryptoService,
     engine: EngineService,
